feat: cycles and modules toggle in settings
