+++ apps/app/components/core/board-view/board-header.tsx
@@ -12,15 +12,13 @@ import {
 // helpers
 import { addSpaceIfCamelCase } from "helpers/string.helper";
 // types
-import { IIssue, NestedKeyOf } from "types";
+import { IIssue } from "types";
 type Props = {
-  provided: DraggableProvided;
   isCollapsed: boolean;
   setIsCollapsed: React.Dispatch<React.SetStateAction<boolean>>;
   groupedByIssues: {
     [key: string]: IIssue[];
   };
-  selectedGroup: NestedKeyOf<IIssue> | null;
   groupTitle: string;
   createdBy: string | null;
   bgColor?: string;
@@ -30,9 +28,7 @@ type Props = {
 export const BoardHeader: React.FC<Props> = ({
   isCollapsed,
   setIsCollapsed,
-  provided,
   groupedByIssues,
-  selectedGroup,
   groupTitle,
   createdBy,
   bgColor,
@@ -44,16 +40,6 @@ export const BoardHeader: React.FC<Props> = ({
     }`}
   >
     <div className={`flex items-center ${!isCollapsed ? "flex-col gap-2" : "gap-1"}`}>
-      <button
-        type="button"
-        {...provided.dragHandleProps}
-        className={`grid h-7 w-7 place-items-center rounded p-1 outline-none duration-300 hover:bg-gray-200 ${
-          !isCollapsed ? "" : "rotate-90"
-        } ${selectedGroup !== "state_detail.name" ? "hidden" : ""}`}
-      >
-        <EllipsisHorizontalIcon className="h-4 w-4 text-gray-600" />
-        <EllipsisHorizontalIcon className="mt-[-0.7rem] h-4 w-4 text-gray-600" />
-      </button>
       <div
         className={`flex cursor-pointer items-center gap-x-1 rounded-md bg-slate-900 px-2 ${
           !isCollapsed ? "mb-2 flex-col gap-y-2 py-2" : ""
