+++ apps/app/components/issues/sidebar.tsx
@@ -56,7 +56,7 @@ type Props = {
 
 const defaultValues: Partial<IIssueLabels> = {
   name: "",
-  colour: "#ff0000",
+  color: "#ff0000",
 };
 
 export const IssueDetailsSidebar: React.FC<Props> = ({
@@ -316,7 +316,7 @@ export const IssueDetailsSidebar: React.FC<Props> = ({
                     >
                       <span
                         className="h-2 w-2 flex-shrink-0 rounded-full"
-                        style={{ backgroundColor: singleLabel?.colour ?? "green" }}
+                        style={{ backgroundColor: singleLabel?.color ?? "green" }}
                       />
                       {singleLabel.name}
                       <XMarkIcon className="h-2 w-2 group-hover:text-red-500" />
@@ -372,7 +372,7 @@ export const IssueDetailsSidebar: React.FC<Props> = ({
                                         >
                                           <span
                                             className="h-2 w-2 flex-shrink-0 rounded-full"
-                                            style={{ backgroundColor: label.colour ?? "green" }}
+                                            style={{ backgroundColor: label.color ?? "green" }}
                                           />
                                           {label.name}
                                         </Listbox.Option>
@@ -422,11 +422,11 @@ export const IssueDetailsSidebar: React.FC<Props> = ({
                       <Popover.Button
                         className={`flex items-center gap-1 rounded-md bg-white p-1 outline-none focus:ring-2 focus:ring-indigo-500`}
                       >
-                        {watch("colour") && watch("colour") !== "" && (
+                        {watch("color") && watch("color") !== "" && (
                           <span
                             className="h-5 w-5 rounded"
                             style={{
-                              backgroundColor: watch("colour") ?? "green",
+                              backgroundColor: watch("color") ?? "green",
                             }}
                           />
                         )}
@@ -444,7 +444,7 @@ export const IssueDetailsSidebar: React.FC<Props> = ({
                       >
                         <Popover.Panel className="absolute right-0 bottom-8 z-10 mt-1 max-w-xs transform px-2 sm:px-0">
                           <Controller
-                            name="colour"
+                            name="color"
                             control={controlLabel}
                             render={({ field: { value, onChange } }) => (
                               <TwitterPicker
