+++ apps/app/components/issues/view-select/due-date.tsx
@@ -1,0 +1,36 @@
+// ui
+import { CustomDatePicker } from "components/ui";
+// helpers
+import { findHowManyDaysLeft } from "helpers/date-time.helper";
+// types
+import { IIssue } from "types";
+
+type Props = {
+  issue: IIssue;
+  partialUpdateIssue: (formData: Partial<IIssue>) => void;
+  isNotAllowed: boolean;
+};
+
+export const ViewDueDateSelect: React.FC<Props> = ({ issue, partialUpdateIssue, isNotAllowed }) => (
+  <div
+    className={`group relative ${
+      issue.target_date === null
+        ? ""
+        : issue.target_date < new Date().toISOString()
+        ? "text-red-600"
+        : findHowManyDaysLeft(issue.target_date) <= 3 && "text-orange-400"
+    }`}
+  >
+    <CustomDatePicker
+      placeholder="N/A"
+      value={issue?.target_date}
+      onChange={(val) =>
+        partialUpdateIssue({
+          target_date: val,
+        })
+      }
+      className={issue?.target_date ? "w-[6.5rem]" : "w-[3rem] text-center"}
+      disabled={isNotAllowed}
+    />
+  </div>
+);
