+++ apps/app/layouts/settings-layout/settings-sidebar.tsx
@@ -1,5 +1,4 @@
 // next
-import { ArrowLeftIcon } from "@heroicons/react/24/outline";
 import Link from "next/link";
 import { useRouter } from "next/router";
